+++ src/app/content/highlights/hooks/locationChange.ts
@@ -6,10 +6,11 @@ import { AnyAction, AppServices, MiddlewareAPI } from '../../../types';
 import { maxHighlightsApiPageSize } from '../../constants';
 import { bookAndPage } from '../../selectors';
 import { Book } from '../../types';
-import { extractDataFromHighlightClientResponse, incrementPage } from '../../utils/sharedHighlightsUtils';
+import { extractDataFromHighlightClientResponse } from '../../utils/sharedHighlightsUtils';
 import { receiveHighlights } from '../actions';
 import * as select from '../selectors';
 import { HighlightData, SummaryHighlightsPagination } from '../types';
+import { incrementPage } from '../utils/paginationUtils';
 
 // TODO - some of this logic could be integrated into src/app/content/highlights/hooks/utils.ts
 // once openstax/rex-web#489 is merged
